fix(Word2Vec): Make nearest() conform to docs
The `Word2Vec.nearest()` function, according to docs, accepts a `max`
parameter which is the maximum number of results returned. This commit
fixes a bug where calling the function with a `max` of 1 would return 0
results. This is due to an `Array.slice()` call that happens -- we were
previously doing `nearestVectors.slice(start,max)` but `start` was
hard-coded to 1. Doing `slice(1,max)` where `max` is 1 returns no
results, so instead we should do `slice(1, max + 1)`.