fix(LSTM): Cast length & temperature options to Number
If you passed the `length` and `temperature` options to `lstm.generate()` as a string (which is easy
to do if you're using vanilla JS and not p5) you could end up with a situation where `generate()`
would interpret `length` as a very large number and freeze up your CPU for a long time. Here I use
the JS trick of prepending a "+" to a variable that could be either a string or a number; this is a
no-op on a Number but it casts a String to a Number, the equivalent of `Number(variable)`.