Add more verbose error to Word2vec.sum & average
In issue #77 it's noted that Word2Vec throws an unhelpful error message
when you ask for the average of 2 words where one word isn't in the
dictionary. This can be traced to a less-than-helpful error message
thrown by `sum()`, which is called by `average()`.

This commit changes `sum()` to throw a "not found" `Error` object that
lists the words that were not found. This object propagates up to the
`average()` function.