feat(KNNImage): Update the KNN Image Classifier to match the dl 0.5
Besides updating the KNN Image Classifier this commit also changes the its api. Now you can
construct the classifier and pass it a optinal HTML videl elt (or in p5 a video.elt) and then you
can call addImageFromVideo() to train it on that specific video and predictFromVideo() to predict
from that video. This allows for the user to create a video of any size and ml5 will handle the
resize. If the video elt is not given, then addImage() and predictFromImage() can still be called
givin image or video as the first argument