refactor(TransformNet): update the TransformNet to dl 0.5
Updating the TransformNet method to dl.05. This commit also changes the name of this method to
FastStyle instead of TransformNet and the predict() methods is now called transfer(). The class has
been redesign so the transfer methods just handles outputing the desire result. Some static methods
have also been addded. The transfer() method now returns a HTMLImage element instead of raw pixels,
so no need to do array3DtoImage in the sketch.