feat(NN and YOLO): removing wip classes and moving them to a wip branch
Removing YOLO and Neural Network class since they are not ready for use yet. YOLO needs some updates
from dl and Neural Network needs to be updated to be more efficient (it was just a toy example). I'm
effectively removing them from this series of commit and moving it to a WIP branch