approachable
I think it's important for us to distinguish between accessibility as "the design of products, devices, services, or environments for people who experience disabilities" and approachability -- meaning friendly materials suitable for beginner audiences. To be clear, accessibility is vital and a priority for this project, but I don't want to be misleading that we have implemented comprehensive accessibility.