Dead link to TM
Hey Guys, 
I'm not quite sure if it works if i file a PR like this, I just found a dead link to teachable machine. It seemed to link to a specific page, which was removed. It could be linking to https://teachablemachine.withgoogle.com/train/audio instead of my proposal of the main page. But going to the homepage would give new users a better understanding where they are, I think!? 
What do you think and if it is not possible to file this PR from here. I can do it the official way as well.